feat: implement file upload hook with drag-and-drop support and validation
