+++ src/components/FileManager/constants.tsx
@@ -22,3 +22,5 @@ export const sidebarWidth = 280;
 export const sidebarTitleDefault = 'Files';
 
 export const BASE_FILE_MANAGER_ICON_SIZE = 20;
+
+export const FILES_DATA_TRANSFER_TYPE = 'Files';
