+++ src/types/file-manager.ts
@@ -1,5 +1,3 @@
-import type { DialFileNodeType } from '@/models/file';
-
 export enum DialFileManagerTabs {
   MyFiles = 'my_files',
   Shared = 'shared',
@@ -13,21 +11,4 @@ export enum DialFileManagerActions {
   Rename = 'rename',
   Download = 'download',
   Delete = 'delete',
-}
-
-export interface DialCopiedItem {
-  sourceUrl: string;
-  destinationUrl: string;
-  overwrite?: boolean;
-  nodeType: DialFileNodeType;
-}
-
-export interface DialDeletedItem {
-  sourceUrl: string;
-  nodeType: DialFileNodeType;
-}
-
-export interface DialUploadFileItem {
-  fileContent: File;
-  name: string;
 }
